requirements.txt
pandas
nltk
torch
torchvision
numpy
Pillow
tqdm
diffusers==0.11.1
transformers
scipy
ftfy
accelerate
